fix: move `usePromptButtons` to correct directory
